Fix: Decorator tests fixed

+++ src/main/java/org/example/pizza/Pizza.java
@@ -5,6 +5,8 @@ import org.example.pizza.AbstractFactoryInterface.IDough;
 import org.example.pizza.AbstractFactoryInterface.IEdge;
 import org.example.user.UserEmployers;
 
+import java.math.BigDecimal;
+
 public abstract class Pizza {
 
     private UserEmployers employers;
@@ -13,18 +15,18 @@ public abstract class Pizza {
     private IDough dough;
     private IEdge edge;
 
-    protected float value;
+    protected BigDecimal value;
 
     protected Pizza(){}
 
-    protected Pizza(float value, String name, IAbstractPizzaFactory pizzaFactory){
+    protected Pizza(BigDecimal value, String name, IAbstractPizzaFactory pizzaFactory){
         this.value = value;
         this.name = name;
         this.dough = pizzaFactory.createDough();
         this.edge = pizzaFactory.createEdge();
     }
 
-    public abstract float getValuePizza();
+    public abstract BigDecimal getValuePizza();
 
     public UserEmployers getEmployers() {
         return employers;
